fixed buttons styling for animation card and moved tap instructions to show only on the animated suitcase component
