+++ .prettierignore
@@ -1,0 +1,6 @@
+# Ignore artifacts:
+build
+coverage
+
+# Ignore all HTML files:
+*.tsx
